move js files to src/js
