+++ src/index.html
@@ -8,6 +8,6 @@
   </style>
 </head>
 <body>
-<script type="module" src="/main.js"></script>
+<script type="module" src="/js/main.js"></script>
 </body>
 </html>
