+++ src/js/lib/index.js
@@ -1,0 +1,1 @@
+export const promisifiedLoad = (loader, file) => new Promise((resolve, reject) => loader.load(file, resolve, null, reject))
